further improvements in the testing shell script
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/opm@197 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ trunk/run_external_tests.sh
@@ -17,19 +17,13 @@
 ################################################################################
 
 
-# This directory must exist, be readable and writeable and must contain a
-# subdirectory called 'tests'
+# It should not be necessary to change anything below this line. See the help
+# message for the possible command-line options.
 #
-TESTDIR=external_tests
+set -eu
 
 
 ################################################################################
-#
-# It should not be necessary to change anything below this line.
-#
-
-
-set -eu
 
 
 # Find the script docu.R (from the pkgutils R package) either in the $PATH or
@@ -71,7 +65,7 @@ format_basename()
 ################################################################################
 
 
-# Run tests based on file comparison.
+# Conduct a certain kind of test based on file comparison.
 #
 do_test()
 {
@@ -95,16 +89,14 @@ do_test()
   if [ $# -eq 0 ] || [ -z "$outfmt" ] || [ -z "$wantedfmt" ] ||
     [ -z "$indir" ] || [ -z "$inext" ] || [ -z "$qdir" ]
   then
-    echo "arguments -d, -f, -im -q and -w must be provided" >&2
+    echo "arguments -d, -f, -i, -q and -w must be provided" >&2
     return 1
   fi
 
+  "$@" "$indir"/*."$inext" 2> "$logfile" || true
+
   local infile wantfile gotfile
   local lastfile=
-
-  mkdir -p "$qdir" || return 1
-
-  "$@" "$indir"/*."$inext" 2> "$logfile" || true
 
   for infile in "$indir"/*."$inext"; do
     wantfile=`format_basename "$infile" "$wantedfmt"`
@@ -135,24 +127,8 @@ do_test()
 ################################################################################
 
 
-num_items()
-{
-  echo $#
-}
-
-
-################################################################################
-
-
-read_opm_version()
-{
-  awk '$1 == "Version:" {print $2; exit}' "$@"
-}
-
-
-################################################################################
-
-
+# Modify the version entry within opm-generated JSON files.
+#
 change_json_version()
 {
   local version=$1
@@ -164,6 +140,8 @@ change_json_version()
 ################################################################################
 
 
+# Modify the version entry within opm-generated YAML files.
+#
 change_yaml_version()
 {
   local version=$1 tmpfile=`mktemp --tmpdir`
@@ -187,40 +165,84 @@ change_yaml_version()
 ################################################################################
 
 
-run_opm=`find_run_opm_script || :`
-if [ "$run_opm" ]; then
-  echo "using script '$run_opm'" >&2
-  echo >&2
-  echo "WARNING: ensure this is the opm version you want to test" >&2
-  echo >&2
-else
-  echo "script 'run_opm.R' not found, exiting now" >&2
+help_msg=
+np=4 # using more cores yielded only little speedup
+run_opm=
+testdir=external_tests
+version=opm_in/DESCRIPTION
+
+OPTIND=1
+while getopts d:hp:s:v: opt; do
+  case $opt in
+    d ) testdir=$OPTARG;;
+    h ) help_msg=yes;;
+    p ) np=$(($OPTARG + 0));;
+    s ) run_opm=$OPTARG;;
+    v ) version=$OPTARG;;
+    * ) exit 1;;
+  esac
+done
+shift $(($OPTIND - 1))
+
+
+if [ "$help_msg" ] || [ $# -gt 0 ]; then
+  cat >&2 <<-__EOF
+	$0 -- test the opm package via its 'run_opm.R' script
+
+	Options:
+	  -d x  Use x as test directory (must contain subdirectory 'tests').
+	  -h    Print this message.
+	  -p x  Use x processors (cores).
+	  -s x  Use x as 'run_opm.R' script.
+	  -v x  Insert opm version x (can also be an R package DESCRIPTION file).
+
+__EOF
   exit 1
 fi
 
 
-if [ $# -eq 0 ]; then
-  version=`read_opm_version opm_in/DESCRIPTION`
+################################################################################
+
+
+[ "$run_opm" ] || run_opm=`find_run_opm_script || :`
+if [ -s "$run_opm" ]; then
+  echo "Using script '$run_opm'..." >&2
+  echo "WARNING: ensure this is the opm version you want to test" >&2
+  echo >&2
 else
-  version=$1
+  echo "script 'run_opm.R' not found and not provided, exiting now" >&2
+  exit 1
 fi
+
+if [ "$version" ]; then
+  [ -s "$version" ] &&
+    version=`awk '$1 == "Version:" {print $2; exit}' "$version"`
+else
+  echo "opm version to insert not found and not provided, exiting now" >&2
+  exit 1
+fi
+
+testfile_dir=$testdir/tests
+if ! [ -d "$testfile_dir" ]; then
+  echo "directory '$testfile_dir' does not exist, exiting now" >&2
+  exit 1
+fi
+
+
+failedfile_dir=$testdir/failed_files # within $testdir, created if necessary
+[ -d "$failedfile_dir" ] && rm -rf "$failedfile_dir"/* ||
+  mkdir "$failedfile_dir"
+errfile=$testdir/tests.err
+outfile=$testdir/tests.out
+rm -f "$errfile" "$outfile"
 
 
 tmpdir=`mktemp --tmpdir -d`
 tmpfile=`mktemp --tmpdir`
 
 
-testfile_dir=$TESTDIR/tests
-failedfile_dir=$TESTDIR/failed_files # within $TESTDIR, created if necessary
-[ -d "$failedfile_dir" ] && rm -rf "$failedfile_dir"/* ||
-  mkdir "$failedfile_dir"
-errfile=$TESTDIR/tests.err
-outfile=$TESTDIR/tests.out
-rm -f "$errfile" "$outfile"
-
-
 # Update the version to let the YAML and JSON tests pass the test irrespective
-# of the actual version. This must later on be set back, see below.
+# of the actual version. This must later on be reversed, see below.
 #
 change_yaml_version "$version" "$testfile_dir"/*.yml
 change_json_version "$version" "$testfile_dir"/*.json
@@ -233,17 +255,17 @@ echo "Testing plot mode..."
 do_test -i csv -d "$testfile_dir" \
   -w "$testfile_dir/%s.ps" -l "$tmpfile" \
   -f "$tmpdir/%s.ps" -q "$failedfile_dir" \
-  Rscript --vanilla "$run_opm" -p 2 -r xyplot -d "$tmpdir" -i '*.csv' \
+  Rscript --vanilla "$run_opm" -p "$np" -r xyplot -d "$tmpdir" -i '*.csv' \
   -k 'TIME:Setup Time,ID' >> "$outfile" &&
     cat "$tmpfile" >> "$errfile"
 
 echo "Testing split mode..."
 # This test only guarantees that if there is nothing to split the original
-# file results
+# file results.
 do_test -i csv -d "$testfile_dir" \
   -w "$testfile_dir"/%s.csv -l "$tmpfile" \
   -f "$tmpdir/%s-00001.csv" -q "$failedfile_dir" \
-  Rscript --vanilla "$run_opm" -p 2 -s , -r split -d "$tmpdir" -i '*.csv' \
+  Rscript --vanilla "$run_opm" -p "$np" -s , -r split -d "$tmpdir" -i '*.csv' \
   -k 'TIME:Setup Time,ID' >> "$outfile" &&
     cat "$tmpfile" >> "$errfile" 
 
@@ -251,7 +273,7 @@ echo "Testing template-collection mode..."
 do_test -i csv -d "$testfile_dir" \
   -w "$testfile_dir/md.template" -l "$tmpfile" \
   -f "$tmpdir/md.template" -q "$failedfile_dir" \
-  Rscript --vanilla "$run_opm" -p 2 -r template -m "$tmpdir/md.template" \
+  Rscript --vanilla "$run_opm" -p "$np" -r template -m "$tmpdir/md.template" \
   -i '*.csv' >> "$outfile" &&
     cat "$tmpfile" >> "$errfile"
 
@@ -259,7 +281,7 @@ echo "Testing template-collection mode with other field separator..."
 do_test -i csv -d "$testfile_dir" \
   -w "$testfile_dir/md.template2" -l "$tmpfile" \
   -f "$tmpdir/md.template2" -q "$failedfile_dir" \
-  Rscript --vanilla "$run_opm" -p 2 -r template -m "$tmpdir/md.template2" \
+  Rscript --vanilla "$run_opm" -p "$np" -r template -m "$tmpdir/md.template2" \
   -s , -i '*.csv' -k 'TIME:Setup Time,ID' >> "$outfile" &&
     cat "$tmpfile" >> "$errfile"
 
@@ -267,7 +289,7 @@ echo "Testing YAML mode..."
 do_test -i csv -d "$testfile_dir" \
   -w "$testfile_dir/%s.yml" -l "$tmpfile" \
   -f "$tmpdir/%s.yml" -q "$failedfile_dir" \
-  Rscript --vanilla "$run_opm" -z -p 2 -a fast -b 0 -r yaml -d "$tmpdir" \
+  Rscript --vanilla "$run_opm" -z -p "$np" -a fast -b 0 -r yaml -d "$tmpdir" \
   -i '*.csv' -k 'TIME:Setup Time,ID' >> "$outfile" &&
     cat "$tmpfile" >> "$errfile"
 
@@ -275,7 +297,7 @@ echo "Testing JSON mode..."
 do_test -i csv -d "$testfile_dir" \
   -w "$testfile_dir/%s.json" -l "$tmpfile" \
   -f "$tmpdir/%s.json" -q "$failedfile_dir" \
-  Rscript --vanilla "$run_opm" -z -p 2 -a smooth -b 0 -r json -d "$tmpdir" \
+  Rscript --vanilla "$run_opm" -z -p "$np" -a smooth -b 0 -r json -d "$tmpdir" \
   -i '*.csv' -k 'TIME:Setup Time,ID' >> "$outfile" &&
     cat "$tmpfile" >> "$errfile"
 
@@ -285,18 +307,18 @@ do_test -i csv -d "$testfile_dir" \
 
 rm -rf "$tmpdir" "$tmpfile"
 
+
 echo
-echo "RESULT:\
-  `grep -F -c '<<<SUCCESS>>>' "$outfile"` successes,\
-  `grep -F -c '<<<FAILURE>>>' "$outfile"` failures,\
-  `grep -F -c '<<<ERROR>>>' "$outfile"` errors"
-echo
-echo "`num_items \`ls "$failedfile_dir"\`` failed files"
+echo -n "RESULT: "
+echo -n "`grep -F -c '<<<SUCCESS>>>' "$outfile"` successes, "
+echo -n "`grep -F -c '<<<FAILURE>>>' "$outfile"` failures, "
+echo -n "`grep -F -c '<<<ERROR>>>' "$outfile"` errors, "
+echo "`ls "$failedfile_dir" | wc -l` quarantined files."
 echo
 
 
 # Fix the version in the YAML and JSON files to avoid SVN updates. Do this in
-# the failed files, too, if any, to avoid annoying reports when manually
+# the quarantined files, too, if any, to avoid annoying reports when manually
 # calling diff.
 #
 change_yaml_version 0.0.0 "$testfile_dir"/*.yml
@@ -306,3 +328,5 @@ change_json_version 0.0.0 "$failedfile_dir"/*.json 2> /dev/null || true
 
 
 ################################################################################
+
+
